+++ plugins/producer_plugin/producer_plugin.cpp
@@ -688,7 +688,7 @@ class producer_plugin_impl : public std::enable_shared_from_this<producer_plugin
                                               || ( !persist_until_expired && _disable_subjective_p2p_billing )
                                               || trx->read_only;
 
-            uint32_t sub_bill = 0;
+            int64_t sub_bill = 0;
             if( !disable_subjective_billing )
                sub_bill = _subjective_billing.get_subjective_bill( first_auth, fc::time_point::now() );
 
@@ -909,7 +909,7 @@ bool producer_plugin::is_producer_key(const chain::public_key_type& key) const
   return false;
 }
 
-uint32_t producer_plugin::get_subjective_bill( const account_name& first_auth, const fc::time_point& now ) const
+int64_t producer_plugin::get_subjective_bill( const account_name& first_auth, const fc::time_point& now ) const
 {
    return my->_subjective_billing.get_subjective_bill( first_auth, now );
 }
@@ -2032,7 +2032,7 @@ bool producer_plugin_impl::process_unapplied_trxs( const fc::time_point& deadlin
             }
 
             // no subjective billing since we are producing or processing persisted trxs
-            const uint32_t sub_bill = 0;
+            const int64_t sub_bill = 0;
             bool disable_subjective_billing = ( _pending_block_mode == pending_block_mode::producing )
                || ( (itr->trx_type == trx_enum_type::persisted) && _disable_subjective_api_billing )
                || ( !(itr->trx_type == trx_enum_type::persisted) && _disable_subjective_p2p_billing )
